Extract TraceConsumer from DDAgentWriter
Unfortunately this required making some things public that were previously private or package visible.  I expect this to be temporary.
